requirements.txt
beautifulsoup4==4.13.4
python-dotenv==1.0.0
bs4==0.0.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==45.0.5
html5lib==1.1
idna==3.10
lxml==6.0.0
Flask==3.0.0
gunicorn==21.2.0
numpy==2.2.4
pandas==2.2.3
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0

pytz==2025.2
requests==2.32.4
six==1.17.0
soupsieve==2.6
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
webencodings==0.5.1

# Web server & runtime
Flask
gunicorn

# Progress bar used by main/tqdm
tqdm

cloudscraper
lxml